requirements.txt
requests
stomper
websocket-client
